00x04

+++ src/effects/AutoDuck.cpp
@@ -301,7 +301,7 @@ bool EffectAutoDuck::Process()
    threshold = threshold * threshold * kRMSWindowSize;
 
    int rmsPos = 0;
-   float rmsSum = 0;
+   double rmsSum = 0;
    // to make the progress bar appear more natural, we first look for all
    // duck regions and apply them all at once afterwards
    std::vector<AutoDuckRegion> regions;
